Parameterize donut chart
